server/requirements.txt
fastapi
uvicorn
fastapi-users
fastapi-users-db-sqlalchemy
sqlalchemy
asyncpg
python-dotenv
python-dateutil